Change error message during signup in cases where a `Person` entry exists in Parfit
If a user tries to sign up but already has a `people.person` entry in parfit gets an error message. Currently it just says `Something went wrong, please try again later`. In this case resetting the password would be the solution.
This affects all users that donated without creating an account or had an old account from the time before auth0 was introduced.

This comit adds a deploy script to change the `auth0-users-validation` error text to something more helpful.

See Auth0 documentation for more information:
https://auth0.com/docs/brand-and-customize/text-customization-new-universal-login/prompt-signup